The couch and Erlang tasks require the environment script.
This is slightly messy. Both invoke the same process so whether you
install Erlang-only or full CouchDB you get the script. Unfortunately,
the decision of where to install (the base BUILD or the custom
COUCH_BUILD) is made independently of the task you invoke and simply
always prefers to install alongside couch, falling back on BUILD which
happens to be where Erlang always goes.